eval/requirements.txt
torch>=2.0.0
transformers>=4.32.0
datasets>=2.12.0
evaluate>=0.4.0
numpy>=1.24.0
pandas>=2.0.0
peft>=0.5.0
tqdm
jinja2
python-dotenv
huggingface_hub